extracted getReceivedCards controller
